ENH: Adding implicitAMI through matrix.solve()
WIP: Attempting to introduce new faces on cyclicAMI and adding
the extra coeff into the matrix through cyclicAMIFvPatchField::manipulateMatrix.

This is a WIP and some variable names will need renaming.

The present implementation matches closely the semi-implicit AMI results
for the tested cases. We need to review the mesh.flux() calculation
which seems to have a problem at the moment.

The lduPrimitiveassembly can assemble mappedPatch or AMIcyclics thorugh
assemble<cyclicAMIPolyPatch> in fvMatrix<Foam::scalar>::solveImplicitCyclic
when the keyWord type  implicitCyclic is found in the linear solver
dictionary.