WIP: First stage to introduce primitive ldu interfaces for parallel Implicit AMI.
This commit is a WIP in the direction to treat AMI as implicit faces
in a assembly matrix. However, faces belonging to AMI which are
processor faces needs to be added to a lduProcessor interface to be
treated as implicitly as possibly. This step is not fully functional
in this commit. Therefore, it is necessary to decompose the case
in such a way that the AMI patches remain in the same processor.