ENH: lagrangian local patch interaction - added 'filter' option
The filter option applies a 'filterFraction' to the number of particles
represented by each parcel when the parcel hits the specified patch(es), i.e.

  mass retained by the patch = filterFraction*nParticles*p.mass

and the number of particles representing the parcel after the interaction
becomes:

  p.nParticles_new = (1 - filterFraction)*p.nParticles_old

The resulting retained mass is added to the 'stick' mass field for further
poat-processing.

Example for the reactingParcelFoam/filter tutorial case where 30% of the
particles are retained by the cyclic patches:

    localInteractionCoeffs
    {
        patches
        (
            "(cyc.*)"
            {
                type        filter;
                filterFraction 0.3;
            }

            walls
            {
                type        rebound;
            }

            "inlet|outlet"
            {
                type escape;
            }
        );