ENH: changed the c coefficients in ISQP and MMA
to Function1, to be able to change their value at runtime.  Useful when
constraints are not feasible at the beginning of the optimisation
(frequent in topology optimisation) and we want to leave some room to
the optimiser to not focus heavily on them at the beginning, but respect
them during the last optimisation cycles.

Has a big impact, especially in MMA, when dealing with volume
constraints in topology optimisation.